+++ engine/language-server/src/main/scala/org/enso/languageserver/data/buffer/Rope.scala
@@ -1,6 +1,7 @@
 package org.enso.languageserver.data.buffer
 import cats.kernel.Monoid
-import org.enso.languageserver.text.{RopeTextEditor, TextEditor}
+import org.enso.languageserver.text.TextEditor
+import org.enso.languageserver.text.editing.RopeTextEditor
 
 /**
   * The measure used for storing strings in the b-tree.
