+++ engine/language-server/src/main/scala/org/enso/languageserver/text/TextEditor.scala
@@ -1,6 +1,6 @@
 package org.enso.languageserver.text
 
-import org.enso.languageserver.text.model.TextEdit
+import org.enso.languageserver.text.editing.model.TextEdit
 
 trait TextEditor[A] {
