+++ engine/language-server/src/main/scala/org/enso/languageserver/text/TextProtocol.scala
@@ -2,7 +2,7 @@ package org.enso.languageserver.text
 
 import org.enso.languageserver.data.{CapabilityRegistration, Client}
 import org.enso.languageserver.filemanager.{FileSystemFailure, Path}
-import org.enso.languageserver.text.model.FileEdit
+import org.enso.languageserver.text.editing.model.FileEdit
 
 object TextProtocol {
