+++ engine/language-server/src/test/scala/org/enso/languageserver/text/editing/RopeTextEditorSpec.scala
@@ -1,8 +1,8 @@
-package org.enso.languageserver.text
+package org.enso.languageserver.text.editing
 
 import org.enso.languageserver.data.buffer.Rope
-import org.enso.languageserver.text.RopeTextEditorSpec._
-import org.enso.languageserver.text.model.{Position, Range, TextEdit}
+import org.enso.languageserver.text.editing.RopeTextEditorSpec._
+import org.enso.languageserver.text.editing.model.{Position, Range, TextEdit}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.must.Matchers
